fix size_t mismatches (#1945)

+++ src/meep.hpp
@@ -2167,7 +2167,7 @@ public:
   std::complex<double> get_field(int c, const vec &loc, bool parallel = true) const;
   std::complex<double> get_field(component c, const vec &loc, bool parallel = true) const;
   double get_field(derived_component c, const vec &loc, bool parallel = true) const;
-  std::vector<int> dft_monitor_size(dft_fields fdft, const volume &where, component c);
+  std::vector<size_t> dft_monitor_size(dft_fields fdft, const volume &where, component c);
   void get_dft_component_dims(dft_chunk **chunklists, int num_chunklists, component c, ivec &min_corner, ivec &max_corner, size_t &array_size, size_t &bufsz, int &rank, direction *ds, size_t *dims, int *array_rank=0, size_t *array_dims=0, direction *array_dirs=0);
 
   // energy_and_flux.cpp
